build: Upgrade to OpenTelemetry 1.5
This change upgrades the OpenTelemetry dependency to version 1.5, which
contains various breaking changes in the metrics API.

+++ opendc-web/opendc-web-runner/src/main/kotlin/org/opendc/web/runner/WebExperimentMonitor.kt
@@ -45,10 +45,10 @@ public class WebExperimentMonitor : ExperimentMonitor {
 
     override fun reportHostSlice(
         time: Long,
-        requestedBurst: Long,
-        grantedBurst: Long,
-        overcommissionedBurst: Long,
-        interferedBurst: Long,
+        requestedBurst: Double,
+        grantedBurst: Double,
+        overcommissionedBurst: Double,
+        interferedBurst: Double,
         cpuUsage: Double,
         cpuDemand: Double,
         powerDraw: Double,
